+++ jme-jwe-scs/jme-jwe-scs-ui/src/app/app-routing-module.ts
@@ -2,10 +2,12 @@ import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 import {ObUnknownRouteModule} from '@oblique/oblique';
 import {Home} from './home/home';
+import {Decrypt} from './decrypt/decrypt';
 
 const routes: Routes = [
 	{path: '', redirectTo: 'home', pathMatch: 'full'},
 	{path: 'home', component: Home},
+	{path: 'decrypt', component: Decrypt},
 	{path: '**', redirectTo: 'unknown-route'}
 ];